refactor: 단순 코드 값 패키지 변경 (#24)
- entity -> enums 로 변경
