+++ RevUp-api/src/main/java/com/revup/question/usecase/UpdateQuestionUseCase.java
@@ -1,6 +1,5 @@
 package com.revup.question.usecase;
 
-import com.revup.question.adaptor.QuestionAdaptor;
 import com.revup.question.dto.request.QuestionUpdateRequest;
 import com.revup.question.dto.request.QuestionAcceptAnswerRequest;
 import com.revup.question.dto.response.QuestionIdResponse;
@@ -12,7 +11,6 @@ import com.revup.question.mapper.QuestionImageMapper;
 import com.revup.question.mapper.QuestionMapper;
 import com.revup.question.service.QuestionService;
 import com.revup.user.entity.User;
-import com.revup.user.util.UserUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
