+++ RevUp-domain/src/main/java/com/revup/answer/enums/AdoptedReview.java
@@ -1,4 +1,4 @@
-package com.revup.answer.entity;
+package com.revup.answer.enums;
 
 import com.revup.answer.exception.InvalidAdoptedReviewException;
