+++ RevUp-domain/src/main/java/com/revup/question/repository/CustomQuestionRepository.java
@@ -2,7 +2,6 @@ package com.revup.question.repository;
 
 import com.revup.question.criteria.QuestionSearchCriteria;
 import com.revup.question.entity.Question;
-import com.revup.question.entity.QuestionType;
 import com.revup.user.entity.User;
 
 import java.time.LocalDateTime;
