+++ RevUp-domain/src/main/java/com/revup/question/repository/CustomQuestionRepositoryImpl.java
@@ -5,8 +5,8 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.revup.common.SkillStack;
 import com.revup.question.criteria.QuestionSearchCriteria;
 import com.revup.question.entity.Question;
-import com.revup.question.entity.QuestionState;
-import com.revup.question.entity.QuestionType;
+import com.revup.question.enums.QuestionState;
+import com.revup.question.enums.QuestionType;
 import com.revup.user.entity.User;
 import jakarta.persistence.LockModeType;
 import lombok.RequiredArgsConstructor;
